Refactor serializers into its own folder
